Fixes #8: Correct confusion around access token and refresh token.
The refresh token is "permanent" and is for sending to the authentication server. The access token is valid for 60 minutes at a time and is for sending to the resource server. Upon expiry the refresh token is used to request a new access token. The access token is the mutable value, and the only time to use the refresh token is when trying to obtain a new access token.